+++ superduper/misc/typing.py
@@ -1,13 +1,9 @@
 import typing as t
 
-from superduper import Component as C
-from superduper.components.datatype import *
-
 File = t.NewType('File', t.AnyStr)
-Blob = t.NewType('Blob', t.Callable)
-Artifact = t.NewType('Artifact', t.Callable)
 Dill = t.NewType('Dill', t.Callable)
 Pickle = t.NewType('Pickle', t.Callable)
 SDict = t.NewType('SDict', t.Dict)
 SList = t.NewType('SList', t.List)
-Component = t.NewType('Component', C)
+JSON = t.NewType('JSON', t.Union[t.Dict, t.List, str, int])
+LeafType = t.NewType('LeafType', t.Any)