+++ test/unittest/component/datatype/test_vector.py
@@ -5,14 +5,6 @@ def test_auto_detect_vector(db):
     db.cfg.auto_schema = True
     db['vectors'].insert([{'x': numpy.random.randn(7)} for _ in range(3)])
 
-    assert 'vector[7]' in db.show('datatype')
+    t = db.load('table', 'vectors')
 
-    schema = next(iter(db.show('schema')))
-    schema = db.load('schema', schema)
-
-    impl = schema.fields['x'].datatype_impl
-
-    assert (
-        impl.__module__ + '.' + impl.__class__.__name__
-        == db.databackend.datatype_presets['vector']
-    )
+    assert t.fields['x'] == 'vector[float64:7]'