+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - No need to add `.signature` to `Model` implementations
 - No need to write `Component.__post_init__` to modify attributes (use `Component.postinit`).
 - Move from in-line encoding to schema-based encoding with `Leaf._fields`
+- No need to define _fields
 
 #### New Features & Functionality